requirements.txt
streamlit
plotly
pandas
numpy
fpdf2
openpyxl
kaleido
